Feat: friendReportPage
