+++ src/pages/FriendStar/FriendReport/FriendReport.tsx
@@ -1,0 +1,11 @@
+
+
+const FriendReport = () => {
+	return (
+		<div>
+			프렌드 리포트
+		</div>
+	)
+}
+
+export default FriendReport;
