+++ src/pages/FriendStar/FriendStar.module.scss
@@ -13,7 +13,7 @@
   top: 0;
   left: 0;
   width: 100dvw;
-  height: 91dvh;
+  height: 100dvh;
   background-size: contain;
   background-repeat: no-repeat;
   z-index: 0;
